requirements.txt
aiomysql==0.1.1
schedule==1.2.1
pyyaml==6.0.1